[Stability][1/1]{check migration disabled in allowed_ptr_locked}
适用范围：{SM8550}
准入id：{OPPO-4436368}
分析：{
sync from qcomm,CR-https://orbit/CR/3291341

migrate_disable is called to prevent moving a task to a different
cpu. In the process it will update the task's cpus_ptr to a
temporary mask to restrict the task to run only on the current
cpu. It has only one cpu set in this new mask. migrate_disable
must always be paired with a migrate_enable call.

When migrate_enable is called it will attempt to manipulate the
allowed ptrs mask just prior to actually enabling migration on
a task. This means __set_cpus_allowed_ptr is called. This calls
a tracehook:

  migrate_enable
  __set_cpus_allowed_ptr
  __set_cpus_allowed_ptr_locked
  trace_android_rvh_set_cpus_allowed_ptr_locked
  android_rvh_set_cpus_allowed_ptr_locked

The purpose of handling this hook in walt, is to prevent the
assignment of a task to a halted cpu. When migrate_enable
restores the affinity to p->cpus_mask and p->cpus_mask consists
of a single cpu and that single cpu is halted, WALT will cause
__set_cpus_allowed_ptr to reject and return early. IOW
p->cpus_ptr is not set to p->cpus_mask.

If the p->cpus_ptr is not restored to point to the p->cpus_mask,
a subsequent migrate_disable/migrate_enable can happen, and
is allowed to happen with the runque lock taken. Since the
p->cpus_ptr is not pointing to the p->cpus_mask, migrate_enable
will incorrectly call __set_cpus_allowed_ptr. This in turn
will take the runque lock again resulting in a spinlock
recursion.

Address this by updating the hook code, checking the
migrate_disabled flag for the task, and allowing the existing
cpu to be used even if that cpu is halted. To achieve this
utilize a different hook that is present in the code already:
trace_android_rvh_set_cpus_allowed_by_task().

}
方案：{NA}
风险及影响[快/稳/省/功能/安全隐私]：{稳}
测试建议：{NA}
跨组依赖(topic name)：{}

Change-Id: I7bfaa86a2e6a6625c88f186c2ec089649bac5c53
Reviewed-on: http://gerrit.scm.adc.com:8080/16868216
almid-check: oppo codebase <codebase@oppo.com>
Reviewed-by: 陈 佳 <chenjia3@oppo.com>
Tested-by: 陈 佳 <chenjia3@oppo.com>
code-check: oppo_codecheck <oppo_codecheck@oppo.com>
Reviewed-by: 党 海蛟-80318865 <danghaijiao@oppo.com>